Fix bug in `add_ref_dcgrid!` function
Instantiate element-conductor lookup tables `arcs_dcgrid_cond`, `convs_ac_cond`, `convs_dc_cond` as empty dictionary if no HVDC component exists in the given power system.

Currently, PowerModelsMCDC errors if an ac-only power system is given, since the abovementioned lookup tables, which are used in the variable declaration, are not instantiated. This is not the expected behaviour.