Simplify `add_ref_dcgrid!`
The previous if/else structure was challenging to maintain. Any new dictionary had to be added in two places, and changes needed to be coordinated in both branches.
With this approach, we create dictionaries for the main DC components at the beginning of the function, in case they don't already exist. This eliminates the need to separately consider the case where those dictionaries don't exist.